engine: java.lang.Integer cannot be cast to java.lang.Long
Java executor service uses long
parameters for specifying the inital
and next execution of a task. When
Integer config values are used directly
in executor service submit method it
raised class cast exception.

Long config values should be used
instead of Integer values.

Bug-Url: https://bugzilla.redhat.com/1491115

+++ backend/manager/modules/bll/src/main/java/org/ovirt/engine/core/bll/AutoStartVmsRunner.java
@@ -82,8 +82,8 @@ public abstract class AutoStartVmsRunner implements BackendService {
     private void init() {
         autoStartVmsToRestart = new CopyOnWriteArraySet<>(getInitialVmsToStart());
 
-        int autoStartVmsRunnerIntervalInSeconds =
-                Config.<Integer>getValue(ConfigValues.AutoStartVmsRunnerIntervalInSeconds);
+        long autoStartVmsRunnerIntervalInSeconds =
+                Config.<Long>getValue(ConfigValues.AutoStartVmsRunnerIntervalInSeconds);
         executor.scheduleWithFixedDelay(
                 this::startFailedAutoStartVms,
                 autoStartVmsRunnerIntervalInSeconds,
